:hammer: Abstract AbstractFileExtImageLoader for multi-platform reuse
